+++ src/main/ruby/core/truffle/ffi/pointer_access.rb
@@ -45,7 +45,7 @@ class Truffle::FFI::Pointer
   alias :read_int8 :read_char
 
   def write_char(value)
-    Truffle.invoke_primitive :pointer_write_char, address, value
+    Truffle.invoke_primitive :pointer_write_char, address, Truffle::Type.rb_to_int(value)
     self
   end
   alias :write_int8 :write_char
@@ -56,7 +56,7 @@ class Truffle::FFI::Pointer
   alias :get_int8 :get_char
 
   def put_char(offset, value)
-    Truffle.invoke_primitive :pointer_write_char, address + offset, value
+    Truffle.invoke_primitive :pointer_write_char, address + offset, Truffle::Type.rb_to_int(value)
     self
   end
   alias :put_int8 :put_char
@@ -70,7 +70,7 @@ class Truffle::FFI::Pointer
 
   def write_array_of_char(ary)
     ary.each_with_index do |value, i|
-      Truffle.invoke_primitive :pointer_write_char, address + (i * 1), value
+      Truffle.invoke_primitive :pointer_write_char, address + (i * 1), Truffle::Type.rb_to_int(value)
     end
     self
   end
@@ -95,7 +95,7 @@ class Truffle::FFI::Pointer
   alias :read_uint8 :read_uchar
 
   def write_uchar(value)
-    Truffle.invoke_primitive :pointer_write_uchar, address, value
+    Truffle.invoke_primitive :pointer_write_uchar, address, Truffle::Type.rb_to_int(value)
     self
   end
   alias :write_uint8 :write_uchar
@@ -106,7 +106,7 @@ class Truffle::FFI::Pointer
   alias :get_uint8 :get_uchar
 
   def put_uchar(offset, value)
-    Truffle.invoke_primitive :pointer_write_uchar, address + offset, value
+    Truffle.invoke_primitive :pointer_write_uchar, address + offset, Truffle::Type.rb_to_int(value)
     self
   end
   alias :put_uint8 :put_uchar
@@ -120,7 +120,7 @@ class Truffle::FFI::Pointer
 
   def write_array_of_uchar(ary)
     ary.each_with_index do |value, i|
-      Truffle.invoke_primitive :pointer_write_uchar, address + (i * 1), value
+      Truffle.invoke_primitive :pointer_write_uchar, address + (i * 1), Truffle::Type.rb_to_int(value)
     end
     self
   end
@@ -145,7 +145,7 @@ class Truffle::FFI::Pointer
   alias :read_int16 :read_short
 
   def write_short(value)
-    Truffle.invoke_primitive :pointer_write_short, address, value
+    Truffle.invoke_primitive :pointer_write_short, address, Truffle::Type.rb_to_int(value)
     self
   end
   alias :write_int16 :write_short
@@ -156,7 +156,7 @@ class Truffle::FFI::Pointer
   alias :get_int16 :get_short
 
   def put_short(offset, value)
-    Truffle.invoke_primitive :pointer_write_short, address + offset, value
+    Truffle.invoke_primitive :pointer_write_short, address + offset, Truffle::Type.rb_to_int(value)
     self
   end
   alias :put_int16 :put_short
@@ -170,7 +170,7 @@ class Truffle::FFI::Pointer
 
   def write_array_of_short(ary)
     ary.each_with_index do |value, i|
-      Truffle.invoke_primitive :pointer_write_short, address + (i * 2), value
+      Truffle.invoke_primitive :pointer_write_short, address + (i * 2), Truffle::Type.rb_to_int(value)
     end
     self
   end
@@ -195,7 +195,7 @@ class Truffle::FFI::Pointer
   alias :read_uint16 :read_ushort
 
   def write_ushort(value)
-    Truffle.invoke_primitive :pointer_write_ushort, address, value
+    Truffle.invoke_primitive :pointer_write_ushort, address, Truffle::Type.rb_to_int(value)
     self
   end
   alias :write_uint16 :write_ushort
@@ -206,7 +206,7 @@ class Truffle::FFI::Pointer
   alias :get_uint16 :get_ushort
 
   def put_ushort(offset, value)
-    Truffle.invoke_primitive :pointer_write_ushort, address + offset, value
+    Truffle.invoke_primitive :pointer_write_ushort, address + offset, Truffle::Type.rb_to_int(value)
     self
   end
   alias :put_uint16 :put_ushort
@@ -220,7 +220,7 @@ class Truffle::FFI::Pointer
 
   def write_array_of_ushort(ary)
     ary.each_with_index do |value, i|
-      Truffle.invoke_primitive :pointer_write_ushort, address + (i * 2), value
+      Truffle.invoke_primitive :pointer_write_ushort, address + (i * 2), Truffle::Type.rb_to_int(value)
     end
     self
   end
@@ -245,7 +245,7 @@ class Truffle::FFI::Pointer
   alias :read_int32 :read_int
 
   def write_int(value)
-    Truffle.invoke_primitive :pointer_write_int, address, value
+    Truffle.invoke_primitive :pointer_write_int, address, Truffle::Type.rb_to_int(value)
     self
   end
   alias :write_int32 :write_int
@@ -256,7 +256,7 @@ class Truffle::FFI::Pointer
   alias :get_int32 :get_int
 
   def put_int(offset, value)
-    Truffle.invoke_primitive :pointer_write_int, address + offset, value
+    Truffle.invoke_primitive :pointer_write_int, address + offset, Truffle::Type.rb_to_int(value)
     self
   end
   alias :put_int32 :put_int
@@ -270,7 +270,7 @@ class Truffle::FFI::Pointer
 
   def write_array_of_int(ary)
     ary.each_with_index do |value, i|
-      Truffle.invoke_primitive :pointer_write_int, address + (i * 4), value
+      Truffle.invoke_primitive :pointer_write_int, address + (i * 4), Truffle::Type.rb_to_int(value)
     end
     self
   end
@@ -295,7 +295,7 @@ class Truffle::FFI::Pointer
   alias :read_uint32 :read_uint
 
   def write_uint(value)
-    Truffle.invoke_primitive :pointer_write_uint, address, value
+    Truffle.invoke_primitive :pointer_write_uint, address, Truffle::Type.rb_to_int(value)
     self
   end
   alias :write_uint32 :write_uint
@@ -306,7 +306,7 @@ class Truffle::FFI::Pointer
   alias :get_uint32 :get_uint
 
   def put_uint(offset, value)
-    Truffle.invoke_primitive :pointer_write_uint, address + offset, value
+    Truffle.invoke_primitive :pointer_write_uint, address + offset, Truffle::Type.rb_to_int(value)
     self
   end
   alias :put_uint32 :put_uint
@@ -320,7 +320,7 @@ class Truffle::FFI::Pointer
 
   def write_array_of_uint(ary)
     ary.each_with_index do |value, i|
-      Truffle.invoke_primitive :pointer_write_uint, address + (i * 4), value
+      Truffle.invoke_primitive :pointer_write_uint, address + (i * 4), Truffle::Type.rb_to_int(value)
     end
     self
   end
@@ -346,7 +346,7 @@ class Truffle::FFI::Pointer
   alias :read_long_long :read_long
 
   def write_long(value)
-    Truffle.invoke_primitive :pointer_write_long, address, value
+    Truffle.invoke_primitive :pointer_write_long, address, Truffle::Type.rb_to_int(value)
     self
   end
   alias :write_int64 :write_long
@@ -359,7 +359,7 @@ class Truffle::FFI::Pointer
   alias :get_long_long :get_long
 
   def put_long(offset, value)
-    Truffle.invoke_primitive :pointer_write_long, address + offset, value
+    Truffle.invoke_primitive :pointer_write_long, address + offset, Truffle::Type.rb_to_int(value)
     self
   end
   alias :put_int64 :put_long
@@ -375,7 +375,7 @@ class Truffle::FFI::Pointer
 
   def write_array_of_long(ary)
     ary.each_with_index do |value, i|
-      Truffle.invoke_primitive :pointer_write_long, address + (i * 8), value
+      Truffle.invoke_primitive :pointer_write_long, address + (i * 8), Truffle::Type.rb_to_int(value)
     end
     self
   end
@@ -404,7 +404,7 @@ class Truffle::FFI::Pointer
   alias :read_ulong_long :read_ulong
 
   def write_ulong(value)
-    Truffle.invoke_primitive :pointer_write_ulong, address, value
+    Truffle.invoke_primitive :pointer_write_ulong, address, Truffle::Type.rb_to_int(value)
     self
   end
   alias :write_uint64 :write_ulong
@@ -417,7 +417,7 @@ class Truffle::FFI::Pointer
   alias :get_ulong_long :get_ulong
 
   def put_ulong(offset, value)
-    Truffle.invoke_primitive :pointer_write_ulong, address + offset, value
+    Truffle.invoke_primitive :pointer_write_ulong, address + offset, Truffle::Type.rb_to_int(value)
     self
   end
   alias :put_uint64 :put_ulong
@@ -433,7 +433,7 @@ class Truffle::FFI::Pointer
 
   def write_array_of_ulong(ary)
     ary.each_with_index do |value, i|
-      Truffle.invoke_primitive :pointer_write_ulong, address + (i * 8), value
+      Truffle.invoke_primitive :pointer_write_ulong, address + (i * 8), Truffle::Type.rb_to_int(value)
     end
     self
   end
@@ -461,7 +461,7 @@ class Truffle::FFI::Pointer
   alias :read_float32 :read_float
 
   def write_float(value)
-    Truffle.invoke_primitive :pointer_write_float, address, value
+    Truffle.invoke_primitive :pointer_write_float, address, Truffle::Type.rb_to_f(value)
     self
   end
   alias :write_float32 :write_float
@@ -472,7 +472,7 @@ class Truffle::FFI::Pointer
   alias :get_float32 :get_float
 
   def put_float(offset, value)
-    Truffle.invoke_primitive :pointer_write_float, address + offset, value
+    Truffle.invoke_primitive :pointer_write_float, address + offset, Truffle::Type.rb_to_f(value)
     self
   end
   alias :put_float32 :put_float
@@ -486,7 +486,7 @@ class Truffle::FFI::Pointer
 
   def write_array_of_float(ary)
     ary.each_with_index do |value, i|
-      Truffle.invoke_primitive :pointer_write_float, address + (i * 4), value
+      Truffle.invoke_primitive :pointer_write_float, address + (i * 4), Truffle::Type.rb_to_f(value)
     end
     self
   end
@@ -511,7 +511,7 @@ class Truffle::FFI::Pointer
   alias :read_float64 :read_double
 
   def write_double(value)
-    Truffle.invoke_primitive :pointer_write_double, address, value
+    Truffle.invoke_primitive :pointer_write_double, address, Truffle::Type.rb_to_f(value)
     self
   end
   alias :write_float64 :write_double
@@ -522,7 +522,7 @@ class Truffle::FFI::Pointer
   alias :get_float64 :get_double
 
   def put_double(offset, value)
-    Truffle.invoke_primitive :pointer_write_double, address + offset, value
+    Truffle.invoke_primitive :pointer_write_double, address + offset, Truffle::Type.rb_to_f(value)
     self
   end
   alias :put_float64 :put_double
@@ -536,7 +536,7 @@ class Truffle::FFI::Pointer
 
   def write_array_of_double(ary)
     ary.each_with_index do |value, i|
-      Truffle.invoke_primitive :pointer_write_double, address + (i * 8), value
+      Truffle.invoke_primitive :pointer_write_double, address + (i * 8), Truffle::Type.rb_to_f(value)
     end
     self
   end